fix int overflow on discovery

+++ internal/core/hosts/discovered_host.go
@@ -7,7 +7,7 @@ type DiscoveredHost struct {
 	HostName                 string            `json:"hostname"`
 	CPUCount                 int               `json:"cpu_count"`
 	SocketCount              int               `json:"socket_count"`
-	TotalMemoryMB            int               `json:"total_memory_mb"`
+	TotalMemoryMB            uint64            `json:"total_memory_mb"`
 	AgentVersion             string            `json:"agent_version"`
 	InstallationSource       string            `json:"installation_source"`
 	FullyQualifiedDomainName *string           `json:"fully_qualified_domain_name,omitempty"`
